requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
tensorflow==2.17.0
pillow==10.2.0
numpy==1.26.4
python-multipart==0.0.9
streamlit==1.32.0
requests==2.31.0
matplotlib==3.8.3